+++ .gitignore
@@ -1,3 +1,3 @@
 /target
 *.dll
-config.toml
+/config.toml
