+++ src/capture/wgc_capture.rs
@@ -12,6 +12,7 @@ use windows::Graphics::DirectX::DirectXPixelFormat;
 use super::ScreenCapture;
 use crate::capture::d3d;
 use crate::capture::yuv_converter::BGR0YUVConverter;
+use crate::config::Config;
 use crate::encoder::FfmpegEncoder;
 use crate::performance_profiler::PerformanceProfiler;
 use crate::result::Result;
@@ -23,10 +24,11 @@ use windows::Win32::Graphics::Direct3D11::{
 };
 
 pub struct WGCScreenCapture<'a> {
-    item: &'a GraphicsCaptureItem,
+    item: GraphicsCaptureItem,
     device: ID3D11Device,
     d3d_context: ID3D11DeviceContext,
     frame_pool: Direct3D11CaptureFramePool,
+    config: &'a Config,
 }
 
 impl<'a> WGCScreenCapture<'a> {
@@ -59,7 +61,7 @@ impl<'a> WGCScreenCapture<'a> {
         Ok((resource, frame))
     }
 
-    pub fn new(item: &'a GraphicsCaptureItem) -> Result<Self> {
+    pub fn new(item: GraphicsCaptureItem, config: &'a Config) -> Result<Self> {
         let item_size = item.Size()?;
         let (device, d3d_device, d3d_context) = d3d::create_direct3d_devices_and_context()?;
         let frame_pool = Direct3D11CaptureFramePool::CreateFreeThreaded(
@@ -73,6 +75,7 @@ impl<'a> WGCScreenCapture<'a> {
             device,
             d3d_context,
             frame_pool,
+            config,
         })
     }
 }
@@ -84,9 +87,8 @@ impl ScreenCapture for WGCScreenCapture<'_> {
         mut encoder: FfmpegEncoder,
         mut output: Box<impl OutputSink + Send + ?Sized>,
         mut profiler: PerformanceProfiler,
-        max_fps: u32,
     ) -> Result<()> {
-        let session = self.frame_pool.CreateCaptureSession(self.item)?;
+        let session = self.frame_pool.CreateCaptureSession(&self.item)?;
 
         let (sender, mut receiver) = tokio::sync::mpsc::channel::<Direct3D11CaptureFrame>(1);
 
@@ -108,7 +110,8 @@ impl ScreenCapture for WGCScreenCapture<'_> {
         let width = self.item.Size()?.Width as u32;
         let use_yuv = false;
         let mut yuv_converter = BGR0YUVConverter::new(width as usize, height as usize);
-        let mut ticker = tokio::time::interval(Duration::from_millis((1000 / max_fps) as u64));
+        let mut ticker =
+            tokio::time::interval(Duration::from_millis((1000 / self.config.max_fps) as u64));
         while let Some(frame) = receiver.recv().await {
             let frame_time = frame.SystemRelativeTime()?.Duration;
             profiler.accept_frame(frame.SystemRelativeTime()?.Duration);
