+++ src/encoder/mod.rs
@@ -1,0 +1,4 @@
+mod ffmpeg;
+mod frame_pool;
+
+pub use ffmpeg::FfmpegEncoder;
